# Conflicts:
#	Dianzhu.ApplicationService/Client/ClientService.cs
#	Dianzhu.CSClient.Presenter/MainPresenter/PIdentyList.cs
#	Dianzhu.HttpApi/App_Code/ORM/ORM003007.cs
#	Ydb.InstantMessage.Tests/Bootstraper.cs
#	Ydb.InstantMessage/Application/ChatService.cs
#	Ydb.Test.Integration/Ydb.Test.Integration.csproj
#	Ydb.Test.Integration/app.config
#	Ydb.Test.Integration/packages.config

+++ Ydb.InstantMessage/Application/ReceptionService.cs
@@ -110,18 +110,33 @@ namespace Ydb.InstantMessage.Application
                 return rsDto;
             }
 
-            for (int i = 0; i < existReceptions.Count; i++)
+            if (existReceptions.Count > 0)
             {
-                receptionRepository.Delete(existReceptions[i]);
+                for (int i = 0; i < existReceptions.Count; i++)
+                {
+                    if (assignCS != existReceptions[i].CustomerServiceId.ToString())
+                    {
+                        receptionRepository.Delete(existReceptions[i]);
+                    }
+                    else
+                    {
+                        rsDto.Id = existReceptions[i].Id;
+                        rsDto.CustomerId = existReceptions[i].CustomerId;
+                        rsDto.CustomerServiceId = existReceptions[i].CustomerServiceId;
+                        rsDto.OrderId = existReceptions[i].OrderId;
+                    }
+                }
             }
+            else
+            {
+                ReceptionStatus status = new ReceptionStatus(customerId, assignCS, string.Empty);
+                receptionRepository.Add(status);
 
-            ReceptionStatus status = new ReceptionStatus(customerId, assignCS, string.Empty);
-            receptionRepository.Add(status);
-
-            rsDto.Id = status.Id;
-            rsDto.CustomerId = customerId;
-            rsDto.CustomerServiceId = assignCS;
-            rsDto.OrderId = status.OrderId;
+                rsDto.Id = status.Id;
+                rsDto.CustomerId = status.CustomerId;
+                rsDto.CustomerServiceId = status.CustomerServiceId;
+                rsDto.OrderId = status.OrderId;
+            }
 
             return rsDto;
         }
